Небольшие изменения структуры файлов
